+++ lua/rustaceanvim/config/internal.lua
@@ -42,11 +42,11 @@ local RustaceanDefaultConfig = {
     ---@class RustaceanHoverActionsConfig
     hover_actions = {
 
-      --- whether to replace Neovim's built-in `vim.lsp.buf.hover`
+      --- whether to replace Neovim's built-in `vim.lsp.buf.hover`.
       ---@type boolean
       replace_builtin_hover = true,
 
-      -- the border that is used for the hover window or explain_error window
+      -- the border that is used for the hover window.
       ---@see vim.api.nvim_open_win()
       ---@type string[][]
       border = {
@@ -60,6 +60,43 @@ local RustaceanDefaultConfig = {
         { '│', 'FloatBorder' },
       },
 
+      --- maximal width of the hover window. Nil means no max.
+      ---@type integer | nil
+      max_width = nil,
+
+      --- maximal height of the hover window. Nil means no max.
+      ---@type integer | nil
+      max_height = nil,
+
+      --- whether the hover window  gets automatically focused
+      --- default: false
+      ---@type boolean
+      auto_focus = false,
+    },
+
+    --- whether use plugin's hover to replace Neovim's built-in `vim.lsp.buf.hover`.
+    ---@type boolean
+    replace_builtin_hover = true,
+
+    --- options same as lsp hover
+    ---@see vim.lsp.util.open_floating_preview
+    ---@class table Options applied to floating windows.
+    float_win_config = {
+
+      -- the border that is used for the hover window or explain_error window
+      ---@see vim.api.nvim_open_win()
+      ---@type string[][] | string
+      border = {
+        { '╭', 'FloatBorder' },
+        { '─', 'FloatBorder' },
+        { '╮', 'FloatBorder' },
+        { '│', 'FloatBorder' },
+        { '╯', 'FloatBorder' },
+        { '─', 'FloatBorder' },
+        { '╰', 'FloatBorder' },
+        { '│', 'FloatBorder' },
+      }, -- maybe: 'double', 'rounded', 'shadow', 'single',
+
       --- maximal width of the hover window or explain_error window. Nil means no max.
       ---@type integer | nil
       max_width = nil,
@@ -68,7 +105,7 @@ local RustaceanDefaultConfig = {
       ---@type integer | nil
       max_height = nil,
 
-      --- whether the hover window or explain_error window gets automatically focused
+      --- whether the window gets automatically focused
       --- default: false
       ---@type boolean
       auto_focus = false,
@@ -301,7 +338,6 @@ local RustaceanDefaultConfig = {
   },
   was_g_rustaceanvim_sourced = vim.g.rustaceanvim ~= nil,
 }
-
 local rustaceanvim = vim.g.rustaceanvim or {}
 local opts = type(rustaceanvim) == 'function' and rustaceanvim() or rustaceanvim
 if opts.tools and opts.tools.executor and type(opts.tools.executor) == 'string' then
@@ -324,6 +360,10 @@ if opts.dap and opts.dap.adapter then
   elseif user_adapter ~= nil then
     RustaceanConfig.dap.adapter = user_adapter
   end
+end
+
+if opts.tools.hover_actions == nil then
+  RustaceanConfig.tools.hover_actions = nil
 end
 
 local check = require('rustaceanvim.config.check')