+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Config: Add `tools.float_win_config` and `replace_builtin_hover` to replace `tools.hover_actions`.
+  Would have to use the `tools.hover_actions` options if they exist.
+
 ## [3.12.2] - 2024-01-07
 
 ### Fixed